OS-6742: "Sandbox" menu section for selected pages
