+++ ngui/ui/src/translations/en-US/app.json
@@ -1615,6 +1615,7 @@
   "s3Prefix": "S3 prefix",
   "safe": "Safe",
   "sampleDataToGiveSenseOfTheProduct": "This is a data sample to give you a sense of the product.",
+  "sandbox": "Sandbox",
   "save": "Save",
   "savePerspectiveTitle": "Save Perspective",
   "savedWithCommitmentsForTheLast30Days": "Saved with commitments for the last 30 days",
