+++ ngui/ui/src/utils/menus/mainMenu.js
@@ -1,5 +1,6 @@
 import React from "react";
 import MainMenuSectionTitle from "components/MainMenuSectionTitle";
+import { MAIN_MENU_SECTION_IDS } from "components/MenuGroupWrapper/reducer";
 import MenuSectionBadge from "components/MenuSectionBadge";
 import anomalies from "./anomaliesMenuItem";
 import cloudCostComparisonMenuItem from "./cloudCostComparisonMenuItem";
@@ -16,6 +17,7 @@ import mlModels from "./mlModelMenuItem";
 import mlRunsets from "./mlRunsetsMenuItem";
 import pools from "./poolsMenuItem";
 import quotas from "./quotasMenuItem";
+import recommendationsArchive from "./recommendationsArchiveMenuItem";
 import recommendations from "./recommendationsMenuItem";
 import resourceLifecycle from "./resourceLifecycleMenuItem";
 import resources from "./resourcesMenuItem";
@@ -23,23 +25,15 @@ import settings from "./settingsMenuItem";
 import taggingPolicies from "./taggingPoliciesMenuItem";
 import users from "./usersMenuItem";
 
-export const MAIN_MENU_SECTION_IDS = Object.freeze({
-  HOME: "home",
-  FINOPS: "finops",
-  ML_OPS: "mlOps",
-  POLICIES: "policies",
-  SYSTEM: "system"
-});
-
 export default [
   {
     id: MAIN_MENU_SECTION_IDS.HOME,
-    items: [home, recommendations, resources, pools, environments]
+    items: [home, recommendations, resources, pools]
   },
   {
     id: MAIN_MENU_SECTION_IDS.FINOPS,
     menuSectionTitle: <MainMenuSectionTitle messageId="finops" />,
-    items: [expenses, expensesMap, finOpsPortal, k8sRightsizing, cloudCostComparisonMenuItem]
+    items: [expenses, expensesMap, finOpsPortal, cloudCostComparisonMenuItem]
   },
   {
     id: MAIN_MENU_SECTION_IDS.ML_OPS,
@@ -51,6 +45,11 @@ export default [
     id: MAIN_MENU_SECTION_IDS.POLICIES,
     menuSectionTitle: <MainMenuSectionTitle messageId="policies" />,
     items: [anomalies, quotas, taggingPolicies, resourceLifecycle]
+  },
+  {
+    id: MAIN_MENU_SECTION_IDS.SANDBOX,
+    menuSectionTitle: <MainMenuSectionTitle messageId="sandbox" />,
+    items: [environments, k8sRightsizing, recommendationsArchive]
   },
   {
     id: MAIN_MENU_SECTION_IDS.SYSTEM,
